Relationship

+++ database/migrations/2019_08_23_061203_create_replies_table.php
@@ -19,7 +19,7 @@ class CreateRepliesTable extends Migration
             $table->text('body');
 
             $table->tinyInteger('user_id')->unsigned();
-            $table->tinyInteger('question_id')->unsigned();
+            $table->unsignedBigInteger('question_id')->unsigned();
 
             $table->foreign('question_id')->references('id')->on('questions')->onDelete('cascade');
